fix(hyprland.conf): do not hide mouse on keypress
It is useful while writing in LibreOffice and on Terminal emulators, but
is annoying when doing graphics like in Canva or GIMP, as there are some
chords with mouse and keyboard and then the mouse disappears. Consider
using hide on keypress functionalities at the client side, Foot terminal
does have such functionality indeed.